fix mac10.6 build
Review URL: https://codereview.chromium.org/26882003

git-svn-id: http://skia.googlecode.com/svn/trunk@11708 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ bench/benchmain.cpp
@@ -654,7 +654,7 @@ int tool_main(int argc, char** argv) {
                     if (frameIntervalTime >= FLAGS_minMs) {
                         frameIntervalComputed = true;
                         loopsPerFrame =
-                            ((double)frameIntervalTotalLoops / frameIntervalTime) * FLAGS_minMs;
+                          (int)(((double)frameIntervalTotalLoops / frameIntervalTime) * FLAGS_minMs);
                         if (loopsPerFrame < 1) {
                             loopsPerFrame = 1;
                         }
